fix: コードレビュー指摘修正 — ヒントメッセージ・重複排除・async化
- registry.rs: `fleet login` → `fleet cp login` ヒントメッセージ修正
- is_sensitive_key を utils.rs に集約（up.rs/deploy.rs の重複排除）
- registry handle_status を async 化（tokio runtime 検出の dead code 解消）
